requirements.txt
fastapi
uvicorn[standard]
pyserial
aiofiles
pydantic
sqlalchemy

